🐛 Fix resume of contentvideo when native hls
When the browser can natively play the hls stream,
the player does not initialize the hls.js library and therefore
`player.getHls()` is `null` which resulted in
a JavaScript error when accessing `player.getHls().startLoad()`
directly and prevented the further execution of the js.

This caused the player to not resume the contentvideo.
This should be fixed now.